automatically register commands in bot_commands
* merges redemption & channel comands into bot_commands
* uses a generic run_command method to call all commands
* automatically registers all bot_commands with the bot
